Add stdin capability
This adds a menu item that pops up a prompt for stdin.
This lets you run example programs and use pdb for example.
Not great UI, but works to make progress.

I'm still now sure how the null return in emscripten works.
It seems to flush and signal EOF, but not documented.

Another thing uncovered: you can write string to stdout that
don't have newlines, this is not handled very cleanly by
the line buffering to call Module['print'] and printErr.