Write more intro
Fix parsing newline issue.
Text/code content now has no trailing newline, it's part of spec not content.
So:
  #%
  stuff
  #%
  more stuff

The 'stuff' part content does not have trailing newline. This is because in spec,
the #% part has to start on a newline. So we consider the newline part of the delimiter,
not the contents.